requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
gunicorn==21.2.0
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.2.0
mysql-connector-python-rf==2.2.2
packaging==23.2
protobuf==4.21.12
pycparser==2.21
PyMySQL==1.1.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
Werkzeug==3.0.1
zipp==3.17.0
